% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.data.3}
\alias{ex.data.3}
\title{Longitudinal data with groups}
\format{A data frame with 600 rows and 4 variables:
\describe{
  \item{id}{id number}
  \item{z}{covariate}
  \item{time}{time variable}
  \item{y}{outcome}
  ...
}}
\usage{
ex.data.3
}
\description{
A dataset used as an example for longRPart2
}
\keyword{datasets}
