% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp2Plot.R
\name{lrp2Plot}
\alias{lrp2Plot}
\title{Longitudinal Recursive Partitioning Plotting Function}
\usage{
lrp2Plot(model)
}
\arguments{
\item{model}{A longrpart2 model.}
}
\description{
Longitudinal Recursive Partitioning Plotting Function
}
\examples{
library(longRPart2)
# example goes here
}
\keyword{effects}
\keyword{longitudinal}
\keyword{mixed}
\keyword{partitioning}
\keyword{recursive}
