% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrpPlot.R
\name{lrpPlot}
\alias{lrpPlot}
\title{Plot Expected Trajectories}
\usage{
lrpPlot(model, smoothing = "n", color = NULL, place = "bottomright")
}
\arguments{
\item{model}{Model object from longRPart2()}

\item{smoothing}{Type of smoothing for trajectoris}

\item{color}{Color to use}

\item{place}{Where to place the plot}
}
\description{
Plot Expected Trajectories
}
