% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrpTreePlot.R
\name{lrpTreePlot}
\alias{lrpTreePlot}
\title{Plot the tree}
\usage{
lrpTreePlot(model, use.n = TRUE, colors = NULL, place = "bottomright")
}
\arguments{
\item{model}{Model object from longRPart2()}

\item{use.n}{Print N in each node}

\item{colors}{Colors to use}

\item{place}{Where to place the plot}
}
\description{
Plot the tree
}
