% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lrp}
\alias{summary.lrp}
\title{Summary results from lrp.}
\usage{
\method{summary}{lrp}(object, ...)
}
\arguments{
\item{object}{An object from lrp.}

\item{...}{Other arguments.}
}
\description{
Summary results from lrp.
}
