% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longRPart2.R
\name{longRPart2}
\alias{longRPart2}
\title{Longitudinal Recursive Partitioning}
\usage{
longRPart2(nlmeFormula, method, rPartFormula, randomFormula, data,
  weight = NULL, R = NULL, control = rpart.control())
}
\arguments{
\item{nlmeFormula}{Formula to be passed to nlme}

\item{method}{Whether to use lme() or nlme(). Use either method="lme" or method="nlme".
This changes what additional arguments need to be passed.}

\item{rPartFormula}{Not sure yet}

\item{randomFormula}{Formula to be passed to random= in lme() or nlme()}

\item{data}{Dataset to be used}

\item{weight}{Sample weights to be passed to rpart}

\item{R}{Correlation matrix to use for nlme. this is correlation=}

\item{control}{Control function to be passed to rpart()}
}
\description{
Longitudinal Recursive Partitioning
}
\examples{
library(longRPart2)
# example goes here
}
\keyword{effects}
\keyword{longitudinal}
\keyword{mixed}
\keyword{partitioning}
\keyword{recursive}

