% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longRPart2.R
\name{longRPart2}
\alias{longRPart2}
\title{Longitudinal Recursive Partitioning}
\usage{
longRPart2(method, nlme.model = NULL, randomFormula, fixedFormula = NULL,
  data, start, rPartFormula, weight = NULL, R = NULL,
  control = rpart.control())
}
\arguments{
\item{method}{Whether to use lme() or nlme(). Use either method="lme" or method="nlme".
This changes what additional arguments need to be passed.}

\item{randomFormula}{Random effects to include for nlme() or lme()}

\item{fixedFormula}{Fixed effects to include for nlme() or lme()}

\item{data}{Dataset}

\item{start}{Starting values for nlme()}

\item{rPartFormula}{Not sure yet}

\item{weight}{Sample weights to be passed to rpart}

\item{R}{Correlation matrix to use for nlme. this is correlation=}

\item{control}{Control function to be passed to rpart()}

\item{model}{Necessary to specify if method="nlme"}
}
\description{
Longitudinal Recursive Partitioning
}
\examples{
library(longRPart2)
# example goes here
}
\keyword{effects}
\keyword{longitudinal}
\keyword{mixed}
\keyword{partitioning}
\keyword{recursive}

