% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp2Plot.R
\name{lrp2Plot}
\alias{lrp2Plot}
\title{Longitudinal Recursive Partitioning Plotting Function}
\usage{
lrp2Plot(model, smooth_method = "gam")
}
\arguments{
\item{model}{A longrpart2 model.}

\item{smooth_method}{Whether to use generalized additive models, smooth_method="gam",
or loess, smooth_method="loess".}
}
\description{
Longitudinal Recursive Partitioning Plotting Function
}
\examples{
library(longRPart2)
}
\keyword{effects}
\keyword{longitudinal}
\keyword{mixed}
\keyword{partitioning}
\keyword{recursive}
