% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lrp}
\alias{plot.lrp}
\title{Plot function for longRPart2}
\usage{
\method{plot}{lrp}(x, ...)
}
\arguments{
\item{x}{A model from lrp.}

\item{...}{Other arguments.}
}
\description{
Plot function for longRPart2
}
